requirements.txt
streamlit
PyMuPDF
langchain==0.1.16
langchain-core==0.1.42
langchain-google-genai==0.0.7
google-generativeai
aiohttp==3.9.3
typing-extensions>=4.7.1
python-dotenv
